% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BKNN-package.R
\docType{package}
\name{BKNN-package}
\alias{BKNN}
\alias{BKNN-package}
\title{BKNN: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wuwill/BKNN}
  \item Report bugs at \url{https://github.com/wuwill/BKNN/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
