% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BKNN.R
\name{get.bnn}
\alias{get.bnn}
\title{Bayesian Nearest Neighours for a data matrix}
\usage{
get.bnn(
  dat,
  sample_info,
  target_sample = sample_info$Name,
  k = 200,
  alpha = 10,
  beta = 10,
  p_gamma = 0.05,
  bnn_name = NULL,
  strata = NULL
)
}
\arguments{
\item{sample_info}{a data.frame for sample information, with one of the columns named as "Name" for sample names or sample IDs, which should be corresponding to \code{data} column names. If sa}

\item{target_sample}{the name(s)/ID(s) of the targeted samples for which BNN posteriors will be calculated}

\item{alpha, beta}{The prior probablities for Beta distributions}

\item{p_gamma}{Proablity of streak extension, ie, not a breakpoint at the next position of the ordered samples}

\item{bnn_name}{either \code{NULL} or a string to name the columns of returned data.frame}

\item{strata}{For stratified weights, provide strata groups for samples. Weights will be calculate per group if provided. Default: NULL, not stratified
\itemize{
\item when \code{strata} is column name of the \code{sample_info} data.frame, the corresponding column will be used as \code{strata}
}}

\item{data}{a matrix to calculate sample distance from, with columns corresponding to samples. Example: principle components for samples}
}
\value{
Bayesian Nearest Neighours results
\itemize{
\item if \code{bnn_name} is \code{NULL}, returns a vector of Bayesian posterior probablities for each sample
\item if \code{bnn_name} is provided as a string, resturns a data.frame, with columns corresponding to the posterior probablities, and the most likely change points (\code{chngpnt}), with columns names as \verb{paste0(c("bnn", "chngpnt), bnn_name)}
}
}
\description{
Bayesian Nearest Neighours for a data matrix
}
