% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BKNN.R
\name{get_sample_weight}
\alias{get_sample_weight}
\title{Get Sample Weights}
\usage{
get_sample_weight(x, strata = NULL)
}
\arguments{
\item{x}{A vector of categorical values, for class types of samples ordered by from the most to the least closest to the target for prediction}

\item{strata}{For stratified weights, provide strata groups for samples. Weights will be calculate per group if provided. Default: NULL, not stratified
\itemize{
\item if \code{NA}, use equal weigths for all non-NA samples
\item if \code{NULL}, calculate weights in all samples without stratification
\item otherwise, calculate weights seprately in each non-NA stratum; The total weights in a stratum is the total number of samples, the total weights of every unique category of are equal to each other.
}}
}
\value{
A numeric vector for weights for each item in the input \code{x}
}
\description{
Get Sample Weights
}
